Split parsimony file
