+++ historydag/parsimony.py
@@ -1,10 +1,7 @@
 """A module implementing Sankoff Algorithm."""
-
 import random
 import ete3
 import numpy as np
-import Bio.Data.IUPACData
-from frozendict import frozendict
 from itertools import product
 from historydag.dag import (
     history_dag_from_histories,
@@ -13,276 +10,7 @@ from historydag.dag import (
     utils,
 )
 from copy import deepcopy
-
-
-class AmbiguityMap:
-    """Implements a bijection between a subset of the power set of an alphabet
-    of bases, and an expanded alphabet of ambiguity codes.
-
-    To look up the set of bases represented by an ambiguity code, use the object like a dictionary.
-    To look up an ambiguity code representing a set of bases, use the ``reversed`` attribute.
-
-    Args:
-        ambiguity_character_map: A mapping from ambiguity codes to collections of bases represented by that code.
-        bases: A collection of bases. If not provided, this will be inferred from the ambiguity_character_map.
-    """
-
-    def __init__(self, ambiguity_character_map, bases=None):
-        ambiguous_values = {
-            char: frozenset(bases) for char, bases in ambiguity_character_map.items()
-        }
-        if bases is None:
-            self.bases = frozenset(
-                base for bset in self.ambiguous_values for base in bset
-            )
-        else:
-            self.bases = frozenset(bases)
-
-        ambiguous_values.update({base: frozenset({base}) for base in self.bases})
-        self.ambiguous_values = frozendict(ambiguous_values)
-        self.reversed = ReversedAmbiguityMap(
-            {bases: char for char, bases in self.ambiguous_values.items()}
-        )
-
-    def __getitem__(self, key):
-        try:
-            return self.ambiguous_values[key]
-        except KeyError:
-            raise KeyError(f"{key} is not a valid ambiguity code for this map.")
-
-    def __iter__(self):
-        return self.ambiguous_values.__iter__()
-
-    def items(self):
-        return self.ambiguous_values.items()
-
-
-class ReversedAmbiguityMap(frozendict):
-    def __getitem__(self, key):
-        try:
-            return super().__getitem__(key)
-        except KeyError:
-            raise KeyError(f"No ambiguity code is defined for the set of bases {key}")
-
-
-_ambiguous_dna_values_gap_as_char = Bio.Data.IUPACData.ambiguous_dna_values.copy()
-_ambiguous_dna_values_gap_as_char.update({"?": "GATC-", "-": "-"})
-_ambiguous_dna_values = Bio.Data.IUPACData.ambiguous_dna_values.copy()
-_ambiguous_dna_values.update({"?": "GATC", "-": "GATC"})
-
-standard_aa_ambiguity_map = AmbiguityMap(_ambiguous_dna_values, "AGCT")
-standard_aa_ambiguity_map_gap_as_char = AmbiguityMap(
-    _ambiguous_dna_values_gap_as_char, "AGCT-"
-)
-
-
-class TransitionAlphabet:
-    def __init__(
-        self, bases=tuple("AGCT"), transition_weights=None, ambiguity_map=None
-    ):
-        self.bases = tuple(bases)
-        self.base_indices = frozendict({base: idx for idx, base in enumerate(bases)})
-        self.yey = np.array(
-            [[i != j for i in range(len(self.bases))] for j in range(len(self.bases))]
-        )
-        if transition_weights is None:
-            self.transition_weights = self.yey
-        else:
-            if len(transition_weights) != len(self.bases) or len(
-                transition_weights[0]
-            ) != len(self.bases):
-                raise ValueError(
-                    "transition_weights must be a nxn matrix, with n=len(bases)"
-                )
-            self.transition_weights = transition_weights
-        if ambiguity_map is None:
-            if self.bases == tuple("AGCT"):
-                self.ambiguity_map = standard_aa_ambiguity_map
-            elif self.bases == tuple("AGCT-"):
-                self.ambiguity_map = standard_aa_ambiguity_map_gap_as_char
-            else:
-                self.ambiguity_map = AmbiguityMap({}, self.bases)
-        else:
-            self.ambiguity_map = ambiguity_map
-
-        # This is applicable even when diagonal entries in transition rate matrix are
-        # nonzero, since it is only a mask on allowable sites based on each base.
-        self.mask_vectors = {
-            code: np.array(
-                [
-                    0 if base in self.ambiguity_map[code] else float("inf")
-                    for base in self.bases
-                ]
-            )
-            for code in self.ambiguity_map
-        }
-
-    def get_ambiguity_from_tuple(self, tup):
-        """Retrieve an ambiguity code encoded by tup, which encodes a set of
-        bases with a tuple of 0's and 1's.
-
-        For example, with ``bases='AGCT'``, ``(0, 1, 1, 1)`` would
-        return `A`.
-        """
-        return self.ambiguity_map.reversed[
-            frozenset(self.bases[i] for i, flag in enumerate(tup) if flag == 0)
-        ]
-
-    def character_distance(self, parent_char, child_char):
-        return self.transition_weights[self.base_indices[parent_char]][
-            self.base_indices[child_char]
-        ]
-
-    def weighted_hamming_distance(self, parent_seq, child_seq):
-        if len(parent_seq) != len(child_seq):
-            raise ValueError("sequence lengths do not match!")
-        return sum(
-            self.character_distance(pchar, cchar)
-            for pchar, cchar in zip(parent_seq, child_seq)
-        )
-
-    def min_character_distance(self, parent_char, child_char):
-        """Allowing child_char to be ambiguous, returns the minimum possible
-        transition weight between parent_char and child_char."""
-        child_set = self.ambiguity_map[child_char]
-        p_idx = self.base_indices[parent_char]
-        return min(
-            self.transition_weights[p_idx][self.base_indices[cbase]]
-            for cbase in child_set
-        )
-
-    def min_weighted_hamming_distance(self, parent_seq, child_seq):
-        """Assuming the child_seq may contain ambiguous characters, returns the
-        minimum possible transition weight between parent_seq and child_seq."""
-        if len(parent_seq) != len(child_seq):
-            raise ValueError("sequence lengths do not match!")
-        return sum(
-            self.min_character_distance(pchar, cchar)
-            for pchar, cchar in zip(parent_seq, child_seq)
-        )
-
-    def get_adjacency_array(self, seq_len):
-        return np.array([self.transition_weights] * seq_len)
-
-    def weighted_hamming_edge_weight(self, field_name):
-        """Returns a function for computing weighted hamming distance between
-        two nodes' sequences.
-
-        Args:
-            field_name: The name of the node label field which contains sequences.
-
-        Returns:
-            A function accepting two :class:`HistoryDagNode` objects ``n1`` and ``n2`` and returning
-            a float: the transition cost from ``n1.label.<field_name>`` to ``n2.label.<field_name>``, or 0 if
-            n1 is the UA node.
-        """
-
-        @utils.access_nodefield_default(field_name, 0)
-        def edge_weight(parent, child):
-            return self.weighted_hamming_distance(parent, child)
-
-        return edge_weight
-
-    def min_weighted_hamming_edge_weight(self, field_name):
-        """Returns a function for computing weighted hamming distance between
-        two nodes' sequences.
-
-        If the child node is a leaf node, and its sequence contains ambiguities, the minimum possible
-        transition cost from parent to child will be returned.
-
-        Args:
-            field_name: The name of the node label field which contains sequences.
-
-        Returns:
-            A function accepting two :class:`HistoryDagNode` objects ``n1`` and ``n2`` and returning
-            a float: the transition cost from ``n1.label.<field_name>`` to ``n2.label.<field_name>``, or 0 if
-            n1 is the UA node.
-        """
-
-        def edge_weight(parent, child):
-            if parent.is_ua_node():
-                return 0
-            elif child.is_leaf():
-                return self.min_weighted_hamming_distance(
-                    getattr(parent.label, field_name), getattr(child.label, field_name)
-                )
-            else:
-                return self.weighted_hamming_distance(
-                    getattr(parent.label, field_name), getattr(child.label, field_name)
-                )
-
-        return edge_weight
-
-
-class UnitTransitionAlphabet(TransitionAlphabet):
-    """A subclass of :class:`TransitionAlphabet` to be used when all
-    transitions have unit cost."""
-
-    def __init__(self, bases="AGCT", ambiguity_map=None):
-        super().__init__(bases=bases, ambiguity_map=None)
-
-    def min_character_distance(self, parent_char, child_char):
-        return int(parent_char not in self.ambiguity_map[child_char])
-
-    def weighted_hamming_distance(self, parent_seq, child_seq):
-        if len(parent_seq) != len(child_seq):
-            raise ValueError("sequence lengths do not match!")
-        return sum(x != y for x, y in zip(parent_seq, child_seq))
-
-
-class VariableTransitionAlphabet(TransitionAlphabet):
-    """A subclass of :class:`TransitionAlphabet` to be used when transition
-    costs depend on site."""
-
-    def __init__(self, bases="AGCT", transition_matrix=None, ambiguity_map=None):
-        assert transition_matrix.shape[1:] == (len(bases), len(bases))
-        super().__init__(bases=bases, ambiguity_map=None)
-        self.transition_weights = None
-        self.sitewise_transition_matrix = transition_matrix
-        self._seq_len
-
-    def character_distance(self, parent_char, child_char, site):
-        return self.sitewise_transition_matrix[site - 1][parent_char][child_char]
-
-    def weighted_hamming_distance(self, parent_seq, child_seq):
-        if len(parent_seq) != self._seq_len or len(child_seq) != self._seq_len:
-            raise ValueError("Sequence lengths must match transition matrix length")
-        return sum(
-            self.character_distance(pchar, cchar, idx + 1)
-            for idx, (pchar, cchar) in enumerate(zip(parent_seq, child_seq))
-        )
-
-    def min_character_distance(self, parent_char, child_char, site):
-        """Allowing child_char to be ambiguous, returns the minimum possible
-        transition weight between parent_char and child_char."""
-        child_set = self.ambiguity_map[child_char]
-        p_idx = self.base_indices[parent_char]
-        return min(
-            self.transition_weights[site - 1][p_idx][self.base_indices[cbase]]
-            for cbase in child_set
-        )
-
-    def min_weighted_hamming_distance(self, parent_seq, child_seq):
-        """Assuming the child_seq may contain ambiguous characters, returns the
-        minimum possible transition weight between parent_seq and child_seq."""
-        if len(parent_seq) != len(child_seq):
-            raise ValueError("sequence lengths do not match!")
-        return sum(
-            self.min_character_distance(pchar, cchar, idx + 1)
-            for idx, (pchar, cchar) in enumerate(zip(parent_seq, child_seq))
-        )
-
-    def get_adjacency_array(self, seq_len):
-        if seq_len != self._seq_len:
-            raise ValueError(
-                f"VariableTransitionAlphabet instance supports sequence length of {self._seq_len}"
-            )
-        return self.sitewise_transition_matrix
-
-
-default_aa_transitions = UnitTransitionAlphabet(bases="AGCT")
-default_aa_gaps_transitions = UnitTransitionAlphabet(bases="AGCT-")
-
+import historydag.parsimony_utils as parsimony_utils
 
 def replace_label_attr(original_label, list_of_replacements={}):
     """Generalizes :meth: ``_replace()`` for namedtuple datatype to replace
@@ -296,7 +24,7 @@ def replace_label_attr(original_label, list_of_replacements={}):
 
 
 def make_weighted_hamming_count_funcs(
-    transition_model=default_aa_transitions,
+    transition_model=parsimony_utils.default_aa_transitions,
     allow_ambiguous_leaves=False,
     sequence_label="sequence",
 ):
@@ -362,7 +90,7 @@ def sankoff_upward(
     node_list,
     seq_len,
     sequence_attr_name="sequence",
-    transition_model=default_aa_transitions,
+    transition_model=parsimony_utils.default_aa_transitions,
     use_internal_node_sequences=False,
 ):
     """Compute Sankoff cost vectors at nodes in a postorder traversal, and
@@ -479,7 +207,7 @@ def sankoff_downward(
     partial_node_list=None,
     sequence_attr_name="sequence",
     compute_cvs=True,
-    transition_model=default_aa_transitions,
+    transition_model=parsimony_utils.default_aa_transitions,
     trim=True,
 ):
     """Assign sequences to internal nodes of dag using a weighted Sankoff
@@ -619,7 +347,7 @@ def disambiguate(
     random_state=None,
     remove_cvs=False,
     adj_dist=False,
-    transition_model=default_aa_transitions,
+    transition_model=parsimony_utils.default_aa_transitions,
     min_ambiguities=False,
 ):
     """Randomly resolve ambiguous bases using a two-pass Sankoff Algorithm on
@@ -773,7 +501,7 @@ def build_trees_from_files(newickfiles, fastafile, **kwargs):
         yield build_tree(newick, fasta_map, **kwargs)
 
 
-def parsimony_score(tree, transition_model=default_aa_transitions):
+def parsimony_score(tree, transition_model=parsimony_utils.default_aa_transitions):
     """returns the parsimony score of a (disambiguated) ete tree.
 
     Tree must have 'sequence' attributes on all nodes.
